==> 2_greg_analysis/4_neuro_astro_cytokines.Rmd <==
---
title: "Seurat Microglia Basics Plotting"
author: "Arpy"
date: '2024-11-12'
output: html_document
---

Neuron and Astrocyte Cytokine/Chemokine Production
Gregory C. 04/2025

```{r load libraries and source, echo = F}
library(tidyverse)
library(Seurat)
library(Libra)
library(patchwork)

source("/Users/chingr/OHSU Dropbox/Saunders Lab's shared workspace/arpy/manuscripts/2023_Thai2P4M_FeigeYoung/ms_analyses/1_karl_analysis/r_functions_paths/_karl_tha2P4M_filepaths.R")
source("/Users/chingr/OHSU Dropbox/Saunders Lab's shared workspace/arpy/manuscripts/2023_Thai2P4M_FeigeYoung/ms_analyses/1_karl_analysis/r_functions_paths/_color_palettes_tha2P4M.R")

cell_de <- read_rds("/Users/chingr/OHSU Dropbox/Saunders Lab's shared workspace/arpy/manuscripts/2023_Thai2P4M_FeigeYoung/ms_analyses/1_karl_analysis/Data/de/DE.tha5.vs.NI.CoarseCellType.rds")

celltype_de <- read_rds("/Users/chingr/OHSU Dropbox/Saunders Lab's shared workspace/arpy/manuscripts/2023_Thai2P4M_FeigeYoung/ms_analyses/1_karl_analysis/Data/de/DE.tha5.vs.NI.CellType.rds")

tha_tha_de <- read_rds("/Users/chingr/OHSU Dropbox/Saunders Lab's shared workspace/arpy/manuscripts/2023_Thai2P4M_FeigeYoung/ms_analyses/1_karl_analysis/Data/de/DE.tha05.vs.tha5.CoarseCellType.rds")

tha_2p4m_de <- read_rds("/Users/chingr/OHSU Dropbox/Saunders Lab's shared workspace/arpy/manuscripts/2023_Thai2P4M_FeigeYoung/ms_analyses/1_karl_analysis/Data/de/DE.tha2p4m.vs.tha05.CoarseCellType.rds")

```

#1 Pulls out DEGs from No Infection vs. Tha5
```{r}
# Subset out just astrocytes and just neurons
astro_de <- subset(cell_de, cell_type == 'Astrocyte')
neuron_de <- subset(cell_de, cell_type == 'Neuron')

astro_tha_de <- subset(tha_tha_de, cell_type == 'Astrocyte')
neuron_tha_de <- subset(tha_tha_de, cell_type == 'Neuron')

astro_2p4m_de <- subset(tha_2p4m_de, cell_type == 'Astrocyte')
neuron_2p4m_de <- subset(tha_2p4m_de, cell_type == 'Neuron')

# Pulls out significant genes
sig_gene_astro <- subset(astro_de, p_val_adj <= 0.01)
sig_gene_neuron <- subset(neuron_de, p_val_adj <= 0.01)

sig_gene_astro_tha <- subset(astro_tha_de, p_val_adj <= 0.01)
sig_gene_neuron_tha <- subset(neuron_tha_de, p_val_adj <= 0.01)

sig_gene_astro_2p4m <- subset(astro_2p4m_de, p_val_adj <= 0.01)
sig_gene_neuron_2p4m <- subset(neuron_2p4m_de, p_val_adj <= 0.01)

```

#2 Isolates Chemokines and Cytokines Only
```{r}
# KEGG Pathway M9809 (Cytokine Cytokine Receptor Interactions)
cytokines <- strsplit('CCL26,CXCL13,CXCR6,TNFSF13B,CCR9,CCL27,EDAR,IL24,TNFRSF13C,IL22RA2,CCR1,CCR3,CCR4,CCR5,CCR6,CCR7,CCR8,CNTF,CNTFR,CSF1,CSF1R,CSF2,CSF2RA,CSF2RB,CSF3,CSF3R,CTF1,IL23R,CX3CR1,IFNLR1,EDA,EGF,EGFR,EPO,EPOR,VEGFD,FLT1,FLT3,FLT3LG,FLT4,TNFRSF13B,CLCF1,IL17RA,AMH,GH1,GH2,AMHR2,GHR,IL17B,TNFRSF21,CCR10,IFNL2,IFNL3,IFNL1,XCR1,CXCR3,CXCL1,CXCL2,CXCL3,IL19,HGF,IFNE,IFNA1,IFNA2,IFNA4,IFNA5,IFNA6,IFNA7,IFNA8,IFNA10,IFNA13,IFNA14,IFNA16,IFNA17,IFNA21,IFNAR1,IFNAR2,IFNB1,IFNG,IFNGR1,IFNGR2,IFNW1,FAS,IL1A,IL1B,IL1R1,IL1RAP,IL2,IL2RA,FASLG,IL2RB,IL2RG,IL3,IL3RA,IL4,IL4R,IL5,IL5RA,IL6,IL6R,IL6ST,IL7,IL7R,CXCL8,CXCR1,IL9,CXCR2,IL9R,IL10,IL10RA,IL10RB,IL11,IL11RA,IL12A,IL12B,IL12RB1,IL12RB2,IL13,IL13RA1,IL15,IL15RA,TNFRSF9,IL17A,IL18,INHBA,INHBB,INHBC,CXCL10,KDR,KIT,CCL4L2,LEP,LEPR,LIF,LIFR,LTA,LTB,LTBR,CCL3L3,MET,KITLG,CXCL9,MPL,NGFR,TNFRSF11B,OSM,IL20,IL21R,IL22,TNFRSF12A,PDGFA,PDGFB,PDGFRA,IL23A,PDGFRB,PF4,PF4V1,IL20RA,IL20RB,PPBP,TNFRSF19,IL17RB,IL26,PDGFC,PRL,PRLR,CCL28,IFNK,CXCL16,IL22RA1,IL21,EDA2R,TNFRSF17,CCL1,CCL2,CCL3,CCL3L3,CCL4,CCL5,CCL7,CCL8,CCL11,CCL13,CCL14,CCL15,CCL16,CCL17,CCL18,CCL19,CCL20,CCL21,CCL22,CCL23,CCL24,CCL25,CXCL6,CXCL11,CXCL5,XCL1,CX3CL1,CXCL12,CRLF2,CXCR5,IL25,BMP2,,BMP7,BMPR1A,BMPR1B,BMPR2,XCL2,TGFB1,TGFB2,TGFB3,TGFBR1,TGFBR2,TNF,TNFRSF1A,TNFRSF1B,TPO,PPBPP1,TNFSF4,CCR2,TNFRSF4,VEGFA,VEGFB,VEGFC,IL1R2,CXCR4,PLEKHO2,GDF5,INHBE,RELT,TSLP,TNFSF11,TNFRSF25,TNFSF14,TNFSF13,TNFSF12,TNFSF10,TNFSF9,TNFRSF14,TNFRSF6B,TNFRSF18,TNFRSF11A,TNFRSF10D,TNFRSF10C,TNFRSF10B,TNFRSF10A,IL18RAP,IL18R1,TNFSF18,ACVR1,ACVR1B,OSMR,ACVR2A,ACVR2B,CD27,ACVRL1,TNFRSF8,TNFSF8,CXCL14,CCL4L2,CD40,CD40LG,CD70,TNFSF15', ",")

cytokines.df <- as.data.frame(cytokines)
names(cytokines.df)[names(cytokines.df) == "c..CCL26....CXCL13....CXCR6....TNFSF13B....CCR9....CCL27....EDAR..." ] <- "cytokines"

# Neurons: 
neuron_cyto <- intersect(sig_gene_neuron$gene,cytokines.df$cytokines)
neuron_cyto_tha <- intersect(sig_gene_neuron_tha$gene,cytokines.df$cytokines)
neuron_cyto_2p4m <- intersect(sig_gene_neuron_2p4m$gene,cytokines.df$cytokines)

# Astrocytes:
astro_cyto <- intersect(sig_gene_astro$gene,cytokines.df$cytokines)
astro_cyto_tha <- intersect(sig_gene_astro_tha$gene,cytokines.df$cytokines)
astro_cyto_2p4m <- intersect(sig_gene_astro_2p4m$gene,cytokines.df$cytokines)

# Plots Relevant Cytokine Genes
plot_astro <- ggplot(data = astro_de, aes(x = avg_logFC, y = -log10(p_val_adj),
                                          label=if_else(
                                          gene %in% astro_cyto,
                                          gene, NA))) + 
        geom_point(color = 'black', size = 2.5) + 
        # geom_hline(yintercept = 0.01, linetype = "dashed", color = "black") +
        xlim(-7, 7) +  # Set x-axis limits
        ylim(-5, 87) +
        geom_point(data = subset(astro_de, gene %in% astro_cyto), color = "red", size = 3) +
        ggrepel::geom_text_repel(size = 3, min.segment.length=0.5, force=4, max.overlaps=80)

plot_neuron <- ggplot(data = neuron_de, aes(x = avg_logFC, y = -log10(p_val_adj),
                                            label=if_else(
                                          gene %in% neuron_cyto,
                                          gene, NA))) + 
    geom_point(color = 'black', size = 2.5) + 
        xlim(-7, 7) +  # Set x-axis limits
        ylim(-5, 87) +
    geom_point(data = subset(neuron_de, gene %in% neuron_cyto), color = "red", size = 3) +
    ggrepel::geom_text_repel(size = 3, min.segment.length=0.5, force=4, max.overlaps=80)

# Tha05 vs Tha5
plot_astro_tha <- ggplot(data = astro_tha_de, aes(x = avg_logFC, y = -log10(p_val_adj),
                                          label=if_else(
                                          gene %in% astro_cyto_tha,
                                          gene, NA))) + 
        geom_point(color = 'gray', size = 2.5) + 
        xlim(-7, 7) +  # Set x-axis limits
        ylim(-5, 87) +
        geom_hline(yintercept = 0.5, linetype = "dashed", color = "black", size = 1.5) + 
        geom_point(data = subset(astro_tha_de, gene %in% astro_cyto_tha), color = "red", size = 3) +
        theme_bw()+ 
        theme(  panel.grid.major = element_line(color = "gray80", size = 0.5),
                panel.border = element_rect(color = "black", fill = NA, size = 0.5),
                 axis.text.x = element_text(size = 25),
                 axis.text.y = element_text(size = 25)) + 
                        labs(x = NULL, y = NULL)
       #  ggrepel::geom_text_repel(size = 6, min.segment.length = 1, force = 20, max.overlaps=80, nudge_x = 2, nudge_y = 15)

plot_neuron_tha <- ggplot(data = neuron_tha_de, aes(x = avg_logFC, y = -log10(p_val_adj),
                                            label=if_else(
                                          gene %in% neuron_cyto_tha,
                                          gene, NA))) + 
    geom_point(color = 'gray', size = 2.5) + 
        xlim(-7, 7) +  # Set x-axis limits
        ylim(-5, 87) +
          geom_hline(yintercept = 0.5, linetype = "dashed", color = "black", size = 1.5) + 
    geom_point(data = subset(neuron_tha_de, gene %in% neuron_cyto_tha), color = "red", size = 3) +
          theme_bw()+ 
        theme(  panel.grid.major = element_line(color = "gray80", size = 0.5),
                panel.border = element_rect(color = "black", fill = NA, size = 0.5),
                 axis.text.x = element_text(size = 25),
                 axis.text.y = element_text(size = 25)) + 
                        labs(x = NULL, y = NULL)
    # ggrepel::geom_text_repel(size = 3, min.segment.length=0.5, force=4, max.overlaps=80)

# Tha05 vs 2p4m
plot_astro_2p4m <- ggplot(data = astro_2p4m_de, aes(x = avg_logFC, y = -log10(p_val_adj),
                                          label=if_else(
                                          gene %in% astro_cyto_2p4m,
                                          gene, NA))) + 
        geom_point(color = 'gray', size = 2.5) + 
        xlim(-7, 7) +  # Set x-axis limits
        ylim(-5, 87) +
          geom_hline(yintercept = 0.5, linetype = "dashed", color = "black", size = 1.5) + 
        geom_point(data = subset(astro_2p4m_de, gene %in% astro_cyto_2p4m), color = "red", size = 3) +
          theme_bw()+ 
        theme(  panel.grid.major = element_line(color = "gray80", size = 0.5),
                panel.border = element_rect(color = "black", fill = NA, size = 0.5),
                 axis.text.x = element_text(size = 25),
                 axis.text.y = element_text(size = 25)) + 
                        labs(x = NULL, y = NULL)
        # ggrepel::geom_text_repel(size = 3, min.segment.length=0.5, force=4, max.overlaps=80)

plot_neuron_2p4m <- ggplot(data = neuron_2p4m_de, aes(x = avg_logFC, y = -log10(p_val_adj),
                                            label=if_else(
                                          gene %in% neuron_cyto_2p4m,
                                          gene, NA))) + 
    geom_point(color = 'gray', size = 2.5) + 
        xlim(-7, 7) +  # Set x-axis limits
        ylim(-5, 87) +
        geom_hline(yintercept = 0.5, linetype = "dashed", color = "black", size = 1.5) +  
    geom_point(data = subset(neuron_2p4m_de, gene %in% neuron_cyto_2p4m), color = "red", size = 3) +
          theme_bw()+ 
        theme(  panel.grid.major = element_line(color = "gray80", size = 0.5),
                panel.border = element_rect(color = "black", fill = NA, size = 0.5),
                 axis.text.x = element_text(size = 25),
                 axis.text.y = element_text(size = 25)) + 
                        labs(x = NULL, y = NULL)
   #  ggrepel::geom_text_repel(size = 3, min.segment.length=0.5, force=4, max.overlaps=80)

# Save Plots

save_path <- "/Users/chingr/OHSU\ Dropbox/Saunders\ Lab\'s\ shared\ workspace/arpy/manuscripts/2023_Thai2P4M_FeigeYoung/ms_analyses/2_greg_analysis/Plots/7_cytokine_analysis/cytokine_volcano/"

 ggsave(paste0(save_path,"astro_tha.png"),
       plot = plot_astro_tha,
       width = 4, height= 4, units='in', 
       dpi = 600, device = "png")
 
  ggsave(paste0(save_path,"neuron_tha.png"),
       plot = plot_neuron_tha,
       width = 4, height= 4, units='in', 
       dpi = 600, device = "png")
  
   ggsave(paste0(save_path,"astro_2p4m.png"),
       plot = plot_astro_2p4m,
       width = 4, height= 4, units='in', 
       dpi = 600, device = "png")
   
    ggsave(paste0(save_path,"neuron_2p4m.png"),
       plot = plot_neuron_2p4m,
       width = 4, height= 4, units='in', 
       dpi = 600, device = "png")

```


#3 Check known receptors that are expressed on microglia in Tha5 condition
```{r}
# Loads NI vs. Tha5 DE
mg_de <- read_rds("/Users/chingr/OHSU Dropbox/Saunders Lab's shared workspace/arpy/manuscripts/2023_Thai2P4M_FeigeYoung/ms_analyses/2_greg_analysis/Data/de/mg_ni_vs_tha5_de.RDS")

# Loads MG cleaned seurat object
mg_so <- read_rds("/Users/chingr/OHSU Dropbox/Saunders Lab's shared workspace/arpy/manuscripts/2023_Thai2P4M_FeigeYoung/ms_analyses/2_greg_analysis/Data/seurat_obj/mgfilt_seurat.RDS")

# known receptors for upregulated cytokines. 
# Commented out genes are not expressed at detectable levels in microglia
FeaturePlot(mg_so, features = c("CCR1"), reduction = "umap")
# FeaturePlot(mg_so, features = c("CCR2"), reduction = "umap")
FeaturePlot(mg_so, features = c("CCR5"), reduction = "umap")
# FeaturePlot(mg_so, features = c("CCR11"), reduction = "umap")
# FeaturePlot(mg_so, features = c("CXCR1"), reduction = "umap")
FeaturePlot(mg_so, features = c("CXCR2"), reduction = "umap")
# FeaturePlot(mg_so, features = c("CXCR3"), reduction = "umap")
FeaturePlot(mg_so, features = c("ACVR1B"), reduction = "umap")
FeaturePlot(mg_so, features = c("ACVR2B"), reduction = "umap")
FeaturePlot(mg_so, features = c("ACVR2A"), reduction = "umap")
FeaturePlot(mg_so, features = c("KIT"), reduction = "umap")

allgenes <- FeaturePlot(mg_so, features = c("CCR1","KIT","CCR5","CXCR2","ACVR1B","ACVR2B","ACVR2A","KIT"), reduction = "umap")

wrap_plots(allgenes)

cyto_genes <- c('CXCL10','CCL3','CCL8')

mg_ligand <- VlnPlot(tha.seurat, features = cyto_genes, group.by = "Experiment", pt.size = 0.1, stack = FALSE)


```

# Global Expression of Cytokines and Receptors
```{r}
# Load full dataset
# tha_seurat <- read_rds("/Users/chingr/OHSU Dropbox/Saunders Lab's shared workspace/arpy/manuscripts/2023_Thai2P4M_FeigeYoung/ms_analyses/1_karl_analysis/Data/seurat_obj/tha.seurat.RDS")

main.path <- "/Users/chingr/OHSU\ Dropbox/Saunders\ Lab\'s\ shared\ workspace/arpy/manuscripts/2023_Thai2P4M_FeigeYoung/ms_analyses/1_karl_analysis/"

tha.summary <- read_rds(paste(main.path,'Data/summaries/tha.celltype.infections.RDS', sep = ''))

tha.genome <- c('Tha_N', 'Tha_P', 'Tha_M', 'Tha_G', 'Tha_L')
tha.seurat <- read_rds(paste(main.path,'Data/seurat_obj/tha.seurat.RDS', sep = ''))
tha.seurat <- AddMetaData(tha.seurat, column_to_rownames(select(tha.summary, CBC, Experiment, infected, CoarseCellType), var='CBC'))

# plot relevant  datasets
cyto_receptor <- c("CCR1","KIT","CCR5","CXCR2","ACVR1B","ACVR2B","ACVR2A","KIT","CXCR3")

allgenes_all_cells <- FeaturePlot(tha.seurat, features = cyto_receptor, reduction = "umap")

wrap_plots(allgenes_all_cells)

vln_receptor <- VlnPlot(tha.seurat, features = c('CCR1','CCR5'), group.by = "CoarseCellType", pt.size = 0.1, stack = FALSE)
vl_test <- VlnPlot(tha.seurat, features = c('CCR1','CCR5','CCR2', 'CCR3'), group.by = "CoarseCellType", pt.size = 0.1, stack = FALSE)


# plot signaling molecules

cyto_ligand <- c("CXCL10","INHBA","KITLG","CCL2","CCL3","CCL4","CCL8","CXCL8","CXCL10","KIT","IL6ST")
cyto_ligand_2 <- c("CCL2", "CCL3","CCL4","CCL8","CCL4L2", "CXCL8", "CXCL10")


allcyto_all_cells <- FeaturePlot(tha.seurat, features = cyto_ligand, reduction = "umap")

wrap_plots(allcyto_all_cells)

vln_ligand <- VlnPlot(tha.seurat, features = cyto_ligand, group.by = "CoarseCellType", pt.size = 0.1, stack = FALSE)

ligand_heatmap <- DoHeatmap(tha.seurat, features = cyto_ligand, group.by = 'CoarseCellType') +
  scale_fill_gradientn(colors = c("white", "blue"))  # Optional color scale''


desired_order <- c("Microglia", "Astrocyte", "Neuron", "hNSC", "Mitotic")
tha.seurat$CoarseCellType <- factor(tha.seurat$CoarseCellType, levels = desired_order)
my_colors = c(Microglia_color, Astrocyte_color, Neuron_color, hNSC_color ,Mitotic_color)

ligand_heatmap_2 <- DoHeatmap(tha.seurat, features = cyto_ligand_2, group.by = 'CoarseCellType') +
  scale_fill_gradientn(colors = c("white", "blue")) +
  
          theme(legend.position = "right",               # keep expression legend
          axis.text.x = element_blank(),           # remove column labels
          axis.ticks.x = element_blank(), 
          plot.title = element_text(size = 26, family = 'Arial', face = "italic", hjust = 0.5, color = "black"),
          axis.title.y = element_text(size = 12),  # Y-axis label
          axis.text.y = element_text(size = 20),
          legend.title = element_text(size = 14),   # Title size
          legend.text = element_text(size = 12)) + 
          labs(x = NULL, y = NULL)

# ggsave(paste0(save_path,"heatmap.png"),
      #  plot = ligand_heatmap_2,
       # width = 12, height= 5, units='in', 
       # dpi = 600, device = "png")

```
### cytokine and receptor heatmaps
```{r}

# LIGANDS
desired_order <- c("Microglia", "Astrocyte", "Neuron", "hNSC", "Mitotic")
tha.seurat$CoarseCellType <- factor(tha.seurat$CoarseCellType, levels = desired_order)
my_colors = c(Microglia_color, Astrocyte_color, Neuron_color, hNSC_color ,Mitotic_color)

all_ligands <- strsplit('ACVRL1, AMH, BMP2, BMP7, CCL1, CCL2, CCL3, CCL3L3, CCL4, CCL4L2, CCL5, CCL7, CCL8, CCL11, CCL13, CCL14, CCL15, CCL16, CCL17, CCL18, CCL19, CCL20, CCL21, CCL22, CCL23, CCL24, CCL25, CCL26, CCL27, CCL28, CLCF1, CNTF, CTF1, CX3CL1, CXCL1, CXCL2, CXCL3, CXCL5, CXCL6, CXCL8, CXCL9, CXCL10, CXCL11, CXCL12, CXCL13, CXCL14, CXCL16, EDA, EDA2R, EGF, FASLG, GDF5, GH1, GH2, HGF, IFNA1, IFNA2, IFNA4, IFNA5, IFNA6, IFNA7, IFNA8, IFNA10, IFNA13, IFNA14, IFNA16, IFNA17, IFNA21, IFNB1, IFNG, IFNK, IFNL1, IFNL2, IFNL3, IFNW1, IL1A, IL1B, IL2, IL3, IL4, IL5, IL6, IL7, IL9, IL10, IL11, IL12A, IL12B, IL13, IL15, IL17A, IL17B, IL18, IL19, IL20, IL21, IL22, IL23A, IL24, IL25, IL26, INHBA, INHBB, INHBC, INHBE, KITLG, LEP, LIF, LTA, LTB, MET, OSM, PDGFA, PDGFB, PDGFC, PF4, PF4V1, PPBP, PRL, TGFB1, TGFB2, TGFB3, TNF, TNFSF4, TNFSF8, TNFSF9, TNFSF10, TNFSF11, TNFSF12, TNFSF13, TNFSF13B, TNFSF14, TNFSF15, TNFSF18, TPO, TSLP, VEGFA, VEGFB, VEGFC, VEGFD, XCL1, XCL2', ', ')

all_receptors <- strsplit('ACVR1, ACVR1B, ACVR2A, ACVR2B, AMHR2, BMPR1A, BMPR1B, BMPR2, CCR1, CCR2, CCR3, CCR4, CCR5, CCR6, CCR7, CCR8, CCR9, CCR10, CD27, CD40, CRLF2, CSF1R, CSF2RA, CSF2RB, CSF3R, CX3CR1, CXCR1, CXCR2, CXCR3, CXCR4, CXCR5, CXCR6, EGFR, EPOR, FAS, GHR, IFNAR1, IFNAR2, IFNGR1, IFNGR2, IFNLR1, IL1R1, IL1R2, IL1RAP, IL2RA, IL2RB, IL2RG, IL4R, IL5RA, IL6R, IL6ST, IL7R, IL9R, IL10RA, IL10RB, IL11RA, IL12RB1, IL12RB2, IL13RA1, IL15RA, IL17RA, IL17RB, IL18R1, IL18RAP, IL20RA, IL20RB, IL21R, IL22RA1, IL22RA2, IL23R, KDR, KIT, LEPR, LIFR, LTBR, MPL, NGFR, OSMR, PDGFRA, PDGFRB, PRLR, TGFBR1, TGFBR2, TNFRSF1A, TNFRSF1B, TNFRSF4, TNFRSF6B, TNFRSF8, TNFRSF9, TNFRSF10A, TNFRSF10B, TNFRSF10C, TNFRSF10D, TNFRSF11A, TNFRSF11B, TNFRSF12A, TNFRSF13B, TNFRSF13C, TNFRSF14, TNFRSF17, TNFRSF18, TNFRSF19, TNFRSF21, XCR1
', ', ')

all_lig <- DoHeatmap(tha.seurat, features = all_ligands[[1]], group.by = 'CoarseCellType', disp.max = 3) +
  scale_fill_gradientn(colors = c("white", "blue")) +
  
          theme(legend.position = "right",               # keep expression legend
          axis.text.x = element_blank(),           # remove column labels
          axis.ticks.x = element_blank(), 
          plot.title = element_text(size = 26, family = 'Arial', face = "italic", hjust = 0.5, color = "black"),
          axis.title.y = element_text(size = 12),  # Y-axis label
          axis.text.y = element_text(size = 13),
          legend.title = element_text(size = 14),   # Title size
          legend.text = element_text(size = 12)) + 
          labs(x = NULL, y = NULL)

# RECEPTORS
desired_order <- c("Microglia", "Astrocyte", "Neuron", "hNSC", "Mitotic")
tha.seurat$CoarseCellType <- factor(tha.seurat$CoarseCellType, levels = desired_order)
my_colors = c(Microglia_color, Astrocyte_color, Neuron_color, hNSC_color ,Mitotic_color)

all_recept <- DoHeatmap(tha.seurat, features = all_receptors[[1]], group.by = 'CoarseCellType', disp.max = 3) +
  scale_fill_gradientn(colors = c("white", "blue")) +
  
          theme(legend.position = "right",               # keep expression legend
          axis.text.x = element_blank(),           # remove column labels
          axis.ticks.x = element_blank(), 
          plot.title = element_text(size = 26, family = 'Arial', face = "italic", hjust = 0.5, color = "black"),
          axis.title.y = element_text(size = 12),  # Y-axis label
          axis.text.y = element_text(size = 13),
          legend.title = element_text(size = 14),   # Title size
          legend.text = element_text(size = 12)) + 
          labs(x = NULL, y = NULL)

ggsave(paste0(save_path ,"ligand_heatmap.png"),
       plot = all_lig,
       width = 6, height= 11, units='in', 
       dpi = 600, device = "png")


ggsave(paste0(save_path ,"receptor_heatmap.png"),
       plot = all_recept,
       width = 6, height= 9, units='in', 
       dpi = 600, device = "png")

```


### CCR1 and CCR5 PLOT
```{r}

# plots CCR1 and CCR5 only acros cell types
desired_order <- c("Microglia", "Astrocyte", "Neuron", "hNSC", "Mitotic")
tha.seurat$CoarseCellType <- factor(tha.seurat$CoarseCellType, levels = desired_order)
my_colors = c(Microglia_color, Astrocyte_color, Neuron_color, hNSC_color ,Mitotic_color)


CCR1_plot <- VlnPlot(tha.seurat, features = c('CCR1'), group.by = "CoarseCellType", pt.size = 0.1, stack = FALSE,  cols = my_colors) +
                        theme(legend.position = "none",
                        plot.title = element_text(size = 26, family = 'Arial', face = "italic", hjust = 0.5, color = "black"),
                        axis.title.x = element_text(),  # X-axis label
                        axis.title.y = element_text(size = 21),  # Y-axis label
                        # axis.text.x = element_text(family = "Arial", face = "bold.italic", size = 18, angle = 90, hjust = 1),   # X-axis tick labels
                        axis.text.x = element_blank(),   # X-axis tick labels
                        axis.text.y = element_text(size = 25),
                        panel.grid.major = element_line(color = "gray80", size = 0.5),
                        panel.border = element_rect(color = "gray50", fill = NA, size = 0.5)) + 
                        labs(x = NULL, y = NULL) +
                        ylim(0,2.5)

CCR5_plot <- VlnPlot(tha.seurat, features = c('CCR5'), group.by = "CoarseCellType", pt.size = 0.1, stack = FALSE,  cols = my_colors) +
                        theme(legend.position = "none",
                        plot.title = element_text(size = 26, family = 'Arial', face = "italic", hjust = 0.5, color = "black"),
                        axis.title.x = element_text(),  # X-axis label
                        axis.title.y = element_text(size = 21),  # Y-axis label
                        # axis.text.x = element_text(family = "Arial", face = "bold.italic", size = 18, angle = 90, hjust = 1),   # X-axis tick labels
                        axis.text.x = element_blank(),   # X-axis tick labels
                        axis.text.y = element_text(size = 25),
                        panel.grid.major = element_line(color = "gray80", size = 0.5),
                        panel.border = element_rect(color = "gray50", fill = NA, size = 0.5)) + 
                        labs(x = NULL, y = NULL) +
                        ylim(0,2.5)

ggsave(paste0(save_path ,"CCR1_cells.png"),
       plot = CCR1_plot,
       width = 3.5, height= 4, units='in', 
       dpi = 600, device = "png")


ggsave(paste0(save_path,"CCR5_cells.png"),
       plot = CCR5_plot,
       width = 3.5, height= 4, units='in', 
       dpi = 600, device = "png")


```

# Microglia Expression of Cytokines
```{r}

require(ggpubr)

save_path <- "/Users/chingr/OHSU\ Dropbox/Saunders\ Lab\'s\ shared\ workspace/arpy/manuscripts/2023_Thai2P4M_FeigeYoung/ms_analyses/2_greg_analysis/Plots/7_cytokine_analysis/microglia_cytokines/"

exp_color <- c(Uninfected_color, Tha_color, Tha2P4M_color, Tha5_color)

select_lig <- c('CCL3', 'CCL4', 'CCL4L2', 'CCL5', 'CCL8',
                                                   'CXCL8', 'CXCL10', 'CXCL16', 'IL1A', 'IL1B', 'IL10', 'IL18',
                                                   'CSF1', 'INHBA')

cytokines.df <- cytokines.df %>% filter(cytokines %in% select_lig)

# for (i in (12)) {
 for (i in (1:length(cytokines.df[[1]]))) {
  if (!(cytokines.df[[1]][[i]] %in% rownames(mg_so))) {
    next
    
  }
  # ANOVA

      df <- data.frame(
        expression = FetchData(mg_so, vars = cytokines.df[[1]][[i]])[,1],
        Experiment = mg_so@meta.data$Experiment
        # Strain = mg_so@meta.data$Strain
      )
      
      df$Experiment <- as.factor(df$Experiment)
      # df$Strain <- as.factor(df$Strain)
      
      anova_result <- aov(expression ~ Experiment, data = df)
      anov_sum <- as.data.frame(summary(anova_result)[[1]])
      
      Tukey <- TukeyHSD(anova_result)
      sig_comparisons <- as.data.frame(Tukey$Experiment)
      sig_comparisons$comparison <- rownames(Tukey$Experiment)
      sig_comparisons <- subset(sig_comparisons, `p adj` < 0.05)
      
      sig_labels <- data.frame(
      group1 = sub("-.*", "", sig_comparisons$comparison),
      group2 = sub(".*-", "", sig_comparisons$comparison),
      p.adj = sig_comparisons$`p adj`,
      y.position = seq(from = max(df$expression) * 1.05, length.out = nrow(sig_comparisons), by = 0.2)  # spread brackets vertically
        )

# Optional: format p-values as stars
sig_labels$label <- ifelse(sig_labels$p.adj < 0.001, "***",
                    ifelse(sig_labels$p.adj < 0.01, "**",
                    ifelse(sig_labels$p.adj < 0.05, "*", "ns")))
      
  # If gene exists, plot
mg_cyto_gene <- VlnPlot(mg_so, features = cytokines.df[[1]][[i]], group.by = "Experiment", pt.size = 0.1, stack = FALSE, add.noise = FALSE, cols = exp_color, alpha = 1) +
                        ylim(0,6) +
                        theme(legend.position = "none",
                        plot.title = element_text(size = 26, family = 'Arial', face = "italic", hjust = 0.5, color = "black"),
                        axis.title.x = element_text(),  # X-axis label
                        axis.title.y = element_text(size = 21),  # Y-axis label
                        # axis.text.x = element_text(family = "Arial", face = "bold.italic", size = 18, angle = 90, hjust = 1),   # X-axis tick labels
                        axis.text.x = element_blank(),   # X-axis tick labels
                        axis.text.y = element_text(size = 25),
                        panel.grid.major = element_line(color = "gray80", size = 0.5),
                        panel.border = element_rect(color = "gray50", fill = NA, size = 0.5)) + 
                        labs(x = NULL, y = NULL) +
                        scale_x_discrete(labels = c("NI_0" = "No Infection", 
                                                    "Tha_05" = "Tha(0.5)", 
                                                    "2P4M_05" = "Tha2P4M(0.5)",
                                                    "Tha_5" = "Tha (5)"))


  mg_cyto_gene$layers[[1]]$aes_params$alpha <- 0.8

# plot <- mg_cyto_gene + stat_pvalue_manual(sig_labels,
#                       xmin = "group1", 
#                       xmax = "group2", 
#                       label = "label", 
#                       y.position = "y.position")

 ggsave(paste0(save_path, cytokines.df[[1]][[i]] ,"_stat.png"),
       plot = mg_cyto_gene,
       width = 3.5, height= 4, units='in', 
       dpi = 600, device = "png")
  
  write.csv(sig_labels, file = paste0(save_path, cytokines.df[[1]][[i]] ,"_stat.csv"), row.names = FALSE)
  write.csv(anov_sum, file = paste0(save_path, cytokines.df[[1]][[i]] ,"_anova_stat.csv"), row.names = FALSE)


}


```



# microglia receptors
```{r}
require(ggpubr)

save_path <- "/Users/chingr/OHSU\ Dropbox/Saunders\ Lab\'s\ shared\ workspace/arpy/manuscripts/2023_Thai2P4M_FeigeYoung/ms_analyses/2_greg_analysis/Plots/7_cytokine_analysis/microglia_cytokines/"

exp_color <- c(Uninfected_color, Tha_color, Tha2P4M_color, Tha5_color)

select_recept <- c('TNFRSF1A', 'CCR1')

cytokines.df <- cytokines.df %>% filter(cytokines %in% select_recept)

# for (i in (12)) {
 for (i in (1:length(cytokines.df[[1]]))) {
  if (!(cytokines.df[[1]][[i]] %in% rownames(mg_so))) {
    next
    
  }
  # ANOVA

      df <- data.frame(
        expression = FetchData(mg_so, vars = cytokines.df[[1]][[i]])[,1],
        Experiment = mg_so@meta.data$Experiment
        # Strain = mg_so@meta.data$Strain
      )
      
      df$Experiment <- as.factor(df$Experiment)
      # df$Strain <- as.factor(df$Strain)
      
      anova_result <- aov(expression ~ Experiment, data = df)
      anov_sum <- as.data.frame(summary(anova_result)[[1]])
      
      Tukey <- TukeyHSD(anova_result)
      sig_comparisons <- as.data.frame(Tukey$Experiment)
      sig_comparisons$comparison <- rownames(Tukey$Experiment)
      sig_comparisons <- subset(sig_comparisons, `p adj` < 0.05)
      
      sig_labels <- data.frame(
      group1 = sub("-.*", "", sig_comparisons$comparison),
      group2 = sub(".*-", "", sig_comparisons$comparison),
      p.adj = sig_comparisons$`p adj`,
      y.position = seq(from = max(df$expression) * 1.05, length.out = nrow(sig_comparisons), by = 0.2)  # spread brackets vertically
        )

# Optional: format p-values as stars
sig_labels$label <- ifelse(sig_labels$p.adj < 0.001, "***",
                    ifelse(sig_labels$p.adj < 0.01, "**",
                    ifelse(sig_labels$p.adj < 0.05, "*", "ns")))
      
  # If gene exists, plot
mg_cyto_gene <- VlnPlot(mg_so, features = cytokines.df[[1]][[i]], group.by = "Experiment", pt.size = 0.1, stack = FALSE, add.noise = FALSE, cols = exp_color, alpha = 1) +
                        ylim(0,4) +
                        theme(legend.position = "none",
                        plot.title = element_text(size = 26, family = 'Arial', face = "italic", hjust = 0.5, color = "black"),
                        axis.title.x = element_text(),  # X-axis label
                        axis.title.y = element_text(size = 21),  # Y-axis label
                        # axis.text.x = element_text(family = "Arial", face = "bold.italic", size = 18, angle = 90, hjust = 1),   # X-axis tick labels
                        axis.text.x = element_blank(),   # X-axis tick labels
                        axis.text.y = element_text(size = 25),
                        panel.grid.major = element_line(color = "gray80", size = 0.5),
                        panel.border = element_rect(color = "gray50", fill = NA, size = 0.5)) + 
                        labs(x = NULL, y = NULL) +
                        scale_x_discrete(labels = c("NI_0" = "No Infection", 
                                                    "Tha_05" = "Tha(0.5)", 
                                                    "2P4M_05" = "Tha2P4M(0.5)",
                                                    "Tha_5" = "Tha (5)"))


  mg_cyto_gene$layers[[1]]$aes_params$alpha <- 0.8

# plot <- mg_cyto_gene + stat_pvalue_manual(sig_labels,
#                       xmin = "group1", 
#                       xmax = "group2", 
#                       label = "label", 
#                       y.position = "y.position")

 ggsave(paste0(save_path, cytokines.df[[1]][[i]] ,"_receptor_specific.png"),
       plot = mg_cyto_gene,
       width = 3.5, height= 4, units='in', 
       dpi = 600, device = "png")
  
  write.csv(sig_labels, file = paste0(save_path, cytokines.df[[1]][[i]] ,"_stat_receptor.csv"), row.names = FALSE)
  write.csv(anov_sum, file = paste0(save_path, cytokines.df[[1]][[i]] ,"_anova_stat_receptor.csv"), row.names = FALSE)


}
```


```{r}



across_cells_CCR <- VlnPlot(tha.seurat, features = c('CCL3','CCL8'), split.by = 'Experiment', group.by = "CoarseCellType", pt.size = 0.1, stack = FALSE)


```
